moved some files around and created a revealer_bot package
